chore(indexer): move logs handlers to `handleLogs.ts`
